split Node class into own source file
